MongoDB and FS integration progress
Made a function that using FS, recursively scans the root directory and all its children, populating the database with the paths of all the found files. Note that the Files schema has been changed to the Paths schema to accomodate both files and folders. For now, I have insured that once the database is initialised, it is immediately populated by the paths of everything in the root directory.