Major Changes to root/router.js
Directories are now traversable using the URL bar, however links for the traversal are yet to be implemented. Work on a /download and /preview route is now underway. Currently, each of the routes will take the path of a desired file as a query. get('/') now redirects to get('/displaydir?='+rootDirectory), which in this case is 'test'.  Also made some minor changes to the display of the date and time. It now displays the actual date and time, rather than milliseconds since Jan 1 1970.

Co-Authored-By: hleni <hleni@users.noreply.github.com>