Loads of changes
Improved the GET /dir to regex match the rest of the link in order to determine the folder to get. Database initialisation also includes folders now. Currently, there is a bug that extensionsless files are considered folders. Chenged the "path" schema to "entry". Includes both files and folders. Included a partial for directories as well. Search now only searches files and is root based (searches everything from the root)