Simplified Schema
Simplified the Schema to Entry. Entries can be both files and folders. Siplifies both initialisation and the display in the dust partials. We now use a singular partial for entries, and dustjs-helpers to separate file and folder display. For now, since database fetches are asynchronous, files and folders are scrambled without any particular order. We will work on improving this.

Co-Authored-By: hleni <hleni@users.noreply.github.com>