Back-end updates
Corrected the POST /dir and created the DELETE /dir. The post only works if we are creating a directory within an existing one (which for the sake of simplicity, is how we envisioned it). The delete only works if the directory we want to delete is completely empty. As per suggestion by Andrea, public/scripts/script.js was moved to project root and renamed to util.js. This commit also saw an attempt to implement skips durig database initialisation (i.e. if file is found during a walk and its already in the database, we simply skip it)