Walk Startup Fixed!
Fixed the startup walk function. If the files are already in the database, then they are not re-saved. Files that are already in the database are skipped. We will also change use of fs to fs-extra, since we require some additional features for folder and file manipulations that are not supported by the regular fs (including callbacks).