Commiting Some Changes to Avoid Merge Conflicts
Changed the route for directory display from GET /dir to GET /dir/display. Fixed the download. The issue was that the fs did not write any data into the files. Also separated the directory creation to PUT /dir/new. Added a hook to said creation to check if the file already exists in the filesystem. Download should work for files not. Next feature to add would be a directory download.