Improved Folder Creation
Folder creation now allows the user to create a folder with a name, even if the name exists. The server will automatically add a "(number)" postfix to the folder name to allow several of them to be created with the same provided name parameter. The method of folder creation has also been changed from PUT /dir to POST /dir to mimic this change in behaviour. The search name parameter in the GET /dir/display is now regex based.