Upload File Route Created
POST /file has been finalised. POSTing files is localised to the current directory the user is in. If a file already exists with the same name, it won't be overwritten, as the name of the file being uploaded will be changed to keep both files in the directory. To achieve this, since using bodyParser does not support multipart/form-data, I decided to use the busboy-body-parser (https://www.npmjs.com/package/busboy-body-parser). On top of the regular support for bodyParser, busboy allso allows for the upload of files in forms, which can be read by the express middleware, and write the buffered data sent by the form both to the fs, as well as the relevant metadata to the db.