Fixed missing code: File Renaming
During the rollback, I accidentally deleted the file renaming function. Now it is fixed to mimic its former use. PUT /file renames a file. If no parameter is specified, default name is set to newname. Renaming the file to a name already existing will add a (number) to the end of the file name.